Extract patching logic
